Rearranged a couple things and updated client.md
