+++ client/src/lib.rs
@@ -11,46 +11,35 @@ extern crate reqwest;
 
 pub mod errors;
 pub mod utils;
-mod methods;
+mod request;
+mod response;
 
-use std::io::Read;
-use cookie::CookieJar;
-use reqwest::{Client, Method, StatusCode, Url};
-use reqwest::header::{Cookie, Headers};
 
+use reqwest::Client;
+use request::Request;
+use response::Response;
 use errors::*;
 
 
-#[derive(Debug, Clone)]
-pub struct Request {
-    pub destination: Url,
-    pub method: Method,
-    pub headers: Headers,
-    pub cookies: CookieJar,
-}
-
-#[derive(Debug, Clone)]
-pub struct Response {
-    pub headers: Headers,
-    pub body: Vec<u8>,
-    pub status: StatusCode,
-}
-
-impl Response {
-    pub fn from_reqwest(original: reqwest::Response) -> Result<Response> {
-        let mut original = original.error_for_status()?;
-        let headers = original.headers().clone();
-        let status = original.status();
-
-        let mut body = Vec::new();
-        original
-            .read_to_end(&mut body)
-            .chain_err(|| "Unable to read the response body")?;
-
-        Ok(Response {
-            status,
-            body,
-            headers,
-        })
+/// Perform a single `GET` request.
+pub fn send_request(req: Request) -> Result<Response> {
+    info!("Sending a GET request to {}", req.destination);
+    if log_enabled!(::log::LogLevel::Debug) {
+        debug!("Sending {} Headers", req.headers.len());
+        for header in req.headers.iter() {
+            debug!("\t{}: {}", header.name(), header.value_string());
+        }
+        for cookie in req.cookies.iter() {
+            debug!("\t{} = {}", cookie.name(), cookie.value());
+        }
     }
+
+    let client = Client::builder()
+        .build()
+        .chain_err(|| "The native TLS backend couldn't be initialized")?;
+
+    client
+        .execute(req.to_reqwest())
+        .chain_err(|| "The request failed")
+        .and_then(|r| Response::from_reqwest(r))
 }
